refactor: remove unused Context
